RAG_ollama/requirements.txt
pandas
openpyxl
scipy
ludwig
textwrap3
requests
pyyaml
langchain
sentence-transformers
chromadb
